Fix compiler warnings and size_t conversions

+++ src/itemsmanagerworker.h
@@ -101,8 +101,8 @@ private:
 	// tabs_signature_ captures <"n", "id"> from JSON tab list, used as consistency check
 	std::vector<std::pair<std::string, std::string> > tabs_signature_;
 	Items items_;
-	int total_completed_, total_needed_;
-	int requests_completed_, requests_needed_;
+    size_t total_completed_, total_needed_;
+    size_t requests_completed_, requests_needed_;
 
 	std::set<std::string> tab_id_index_;
 	std::string tabs_as_string_;
